+++ pkg/analyzer/lib/src/dart/ast/ast.dart
@@ -2019,11 +2019,8 @@ class CompilationUnitImpl extends AstNodeImpl implements CompilationUnit {
   @override
   LineInfo lineInfo;
 
-  /// The major component of the actual language version (not just override).
-  int languageVersionMajor;
-
-  /// The minor component of the actual language version (not just override).
-  int languageVersionMinor;
+  /// The language version information.
+  LibraryLanguageVersion languageVersion;
 
   @override
   final FeatureSet featureSet;